+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 30.33.1
+
+# Android Client:
+* Backported API for compatibility with 31.0.0
+
+
 ## 30.33.0
 
 # Android Client: